fixed file closure bug and name clash bug
The file closure bug was that, even after the video capture was
complete, the video still could not be opened in a video player,
until the program quit. Presumably this was due to the file not
actually being closed. Maybe there's a bug in SharpAvi. Opening
the file as a FileStream as a separate initial step, and
constructing the AviWriter from the FileStream, fixes this, even
when the AviWriter is responsible for closing the FileStream.

The name clash bug was that, if the filename to write existed, the
program crashed. This would happen anytime two separate video
captures were done during the same second, due to the ay the files
are named. The fix for the file closure bug by itself (not
committed) transformed the behavior here into silently appending
data to the existing video stream in the existing file, which is
still not what I want. Now, the file is only opened for creation;
if opening the file fails because the file doesn't exist, an
alternate name is used. Up to int.MaxValue names per second are
supported (lol), so I'm not too worried about the range.