turned off autoscaling
This program uses a fixed-size (800x600) canvas, which should not
itself be scaled or changed in size with different desktop scaling
values, so DPI-based scaling is unsuitable.

Most applications should use font-based scaling but that would be
unsuitable here as well, since the canvas PictureBox is the only
control shown in the client area of the window. The UI should look
the same, as much as possible, at any display scaling.