show a toast notification for each capture
This is not the way I'm going to want it to work, though:

(1) Clicking the toast should bring up the containing folder in
a file browser (explorer) window, with the video file selected.
Instead, it does nothing if the program is running, and it opens
the program if it is not running.

(2) Undismissed toasts should either be kept or discarded on exit.
Right now some unholy combination of these behaviors is apparently
occurring. [EDIT: No, I completely forgot to call Uninstall().]

(3) Multiple captures should probably be consolidated into a single
toast (when one or more previous captures' toasts, from the _same
run_ of the program, are undismissed), with that single toast
reshown each time. Assuming this is feasible, which I believe it
is.

(4) In a toast, all of the path except the filename should be
omitted if the file is in (what is still) the user's Videos
directory.