fade/unfade H.264 menu item based on x264vfw availability
This fades and disables the item for H.264 when it can't be used
and unfades and enables it when it can. It also falls back to
Motion JPEG if H.264 becomes unavailable when it is already
selected. The check happens every time the menu is opened, and also
when recording is about to start (to avoid a crash if x264vfw is
uninstalled while VidDraw is running, followed by recording,
without the system/winow menu being opened in between).

I've also done some code cleanup, mainly:

- No need for my own MF type, CsWin32 provides MENU_FLAGS.

- Removed byte-order mark from all tracked files.