fixed saving/loading custom colors
Reading ColorDialog.CustomColors creates an array, so modifying it
doens't help. Handily, assigning an array does exactly what I want:
if it's too short, it's padded, and if it's too long, it's cut off.

Adding a new custom color often overwrites an existing one, even
when I don't make an attempt to do that. I'm not sure if this is a
problem with ColorDialog, or maybe I don't know how to use it
properly, or maybe it is a bug in my program.