naively quoted the explorer select arg
Unlike other operating systems, on Windows, paths cannot contain
double-quote characters--or at least on filesystems Windows usually
uses, like NTFS, they can't. So just putting "" around the argument
is probably better than nothing.

What I'm not sure about is if a path could end in a character that
would be taken as escaping the closing double-quote character. (A
path containing a double quote, escaped or not, would presumably be
an invalid path, but even if that is always so, not all arguments,
in general, are paths, so there may still be an escaping mechanism,
using ^ or \ or something else, that could cause trouble, I fear.)