synchronize toast cleanup per-architecture; show architecture in title bar
This fixes a regression introduced in 1e16522 where running 32- and
64-bit builds at the same time could cause toasts from one not to
be cleaned up.

Now, toasts *for an architecture* are cleaned up when an exiting
VidDraw process *of that architecture* detects that it is the last
VidDraw process *of that architecture*.

This could give nonintuitive results, if a user runs two or more
VidDraw processes of one architecture and one or more VidDraw
processes of the other architecture at the same time and doesn't
recognize the architecture distinction as important. So I've also
added the architecture to the main window's title bar text. This is
really something I should've done anyway, ever since using a codec
(x264vfw) that can be installed for one architecture and not the
other (and that is separately confgured for each architecture).

I'm still synchronizing the checks with the same mutex, which means
there remains a strange dependency between x64 and x86 VidDraw
processes. This is intentional, because I'm not sure if I'm going
to keep doing cleanup in this somewhat undesirable per-architecture
way. If I don't, I may be glad, for the purpose of compatibility
between older and newer versions of VidDraw, that I kept the "main"
mutex architecture-neutral (and kept its name the same).