unsuccessful attempt to "degrade gracefully" on non-toasting OSes
This manages only to defer the crash a bit, because of the two
exceptions that are thrown, nothing is achieved by attempting to
catch EntryPointNotFoundException. It is thrown (at least a little)
later, during cleanup.

It looks like EAFP is not going to work here, and LBYL will be
needed. Fortunately, which operating system the program is running
on does not change during a single run of the program, so that's no
problem... so long as I do the check correctly.