non-toasting OSes "degrade gracefully" to immediate open+select
This is achieved by checking before performing operations related
to toast notifications, and if we're not on Windows 10, or our
version of Windows 10 is not high enough, either doing an
alternative action or nothing, as appropriate.

Note: As of commit time, this must not be merged to master, since
it has *only* been tested on a Windows 7 system. This could very
well have totally broken toast notifications on systems that do
support them. I still need to test that.