SpecialFolder.SystemX86 does work on 32-bit OSes (tested)
Though I wonder if it's even necessary. Is a 32-bit process
accessing files under the path of a 64-bit system folder guaranteed
to be mapped to the corresponding path under the 32-bit system
folder, even on systems where those folders' paths are unusual (and
unusual in different ways from one another)?

Btw, the previous commit (c028e6f) apears not to have broken
anything relating to toast notifications on systems that do support
them, so this branch (compat) can likely soon be merged to master.