added license and started on documentation
The license shown won't ultimately apply to every file in the
repository, since I plan to keep a few dependencies in the repo:
a font or two, and maybe some CSS or JavaScript libraries. These
will be for the "About" browser (really, a help browser with
reduced expectations). I haven't even started on that, and no
dependencies are checked in yet, so it's not a problem that the
LICENSE.md file that README.md says will explain exactly what is
licensed what way does not exist yet.

README.md, which the "about" file will be to some (hopefully large
extent) based on, is itself not done yet.