started on html help
The HTML help, which will be shown in the (forthcoming) built-in
help browser when the user clicks the About VidDraw... menu item,
is generated from README.md with pandoc. A template specifies the
rest of the page, including the <head> element, and the stylesheet
and dependencies are provided along with it.

At the moment I'm pulling in a few dependencies by CDN. This is
just so I can develop and test incrementally. They won't be gotten
by CDN in the finished HTML help (though I might put it back that
way later in a separate gh-pages branch if I decide to host the
help as its own "site"). This is because:

(1) The HTML help should work just as well when the user is
    offline.

(2) Milligram's very strange root font size (set so that one can,
    for example, say 1.8rem to mean 18px) means that if Milligram
    is not available, page elements will display at wildly wrong
    sizes.

(3) There is a potential privacy issue, in that users expect off-
    site resources to be accessed when browsing the web, and users
    expect dependencies may be retrieved when they install, build,
    or *first run* a program, but users don't expect programs to
    phone home--or in this case, to phone somewhere else--every
    time they are run, without a good reason and disclosure.

Eventually, index.html will be checked into source control, so
I don't have to make running pandoc an essential part of the build
process. (If I did, users who clone the repo and run "dotnet run"
would get an error if they didn't have pandoc. I want people to be
able to build this without having pandoc.)