no hyphens in list (works around IE display bug)
IE disappears the SharpAvi.Net5 list item in the dependencies list
if hyphenation is specifically disabled in anchors in the list. But
it looks really bad for small bracketed links to be hyphenated (and
it was hyphenating "[nuget]" as "[n-uget]"), so better to just not
use hyphens there at all. The nature of this list is that hyphens
are unlikely to make it any more compact, anyway.