give h2 lower weight but more letterspacing
At a font weight of 600, it dominated compared to the much thinner
h1. At a font weight of 400, the heavier h3 was not obviously a
subheading. So I've reduced its font weight to 400 but given it
more letterspacing. Together with their preexisting bottom borders,
I think this makes the h2 headings look important enough that the
h3 and h4 can stay at weight 600, while still not overwhelming the
h1.