moved/renamed html from doc/index.html to README.html (in root)
Correspondingly renamed the doc directory to doc-assets.

Having README.md and the HTML file generated from it in the same
location makes so I don't have to alter paths in href attributes.

It also makes sense that files conceptually representing the same
document (as they will, even once a flexgrid or floating ToC is
added via template.html5 to README.html) go in the same place.

Just as index.html wasn't tracked, I'm not ready to commit the
generated README.html (as it is so rapidly changing, and changes to
it, if committed, may distract from changes in the files involved
in generating it). But, as before, I do plan to commit and track
README.html eventually, so users don't need to have and run pandoc.