use local fonts and js libs, updating versions; add links and notices
The README change is not finished. Some of the licenses don't
display correctly (and arguably none display well). The notices
should be part of the generated README.html (or whatever it will
end up being called) but they should be split out of README.md
because GitHub won't apply a max-height to the pre element with
the Apache license (plus, it's nicer to have them as a separate
file anyway, I think, even if no individual licenses were long).
See TODO and FIXME comments in README.md.

Maybe the notices should show in a separate tab in VidDraw's
(forthcoming) built-in help browser, too.

This commit makes loading, reloading, and perhaps some other
actions very slow in Internet Explorer. I'm hoping they will not be
slow in the WebBrowser control (which embeds IE but doesn't carry
its full interface and functionality). My guess is that this has to
do with the licenses, but it might also have to do with CSS that IE
doesn't process efficiently. It would be worthwhile to look at
cases where hover and focus cause a border to exist sometimes and
not other times, or to have a different thickness (border-width)
sometimes from other times. I suspect this might be slower than
if it always had a border of the same thickness but it was
transparent when not used. Unfortunately, in one case, I am using
a size difference, which there might not be a better way to
achieve. (The way I'm doing it, I don't think the page should have
to be laid out again, but I don't know.) My guess is actually that
this has nothing to do with it, but I don't want to forget about
this in case it is.

This commit enables (good) kbd styling, which wasn't enabled before.