decreased right-side padding
The idea here is that I have so much padding on the left side for
AnchorJS links to appear, but I don't need that much space on the
right (and I made the Fork me on GitHub banner translucent so the
only place that might need it doesn't).

But this asymmetry looks bad. Or maybe just having that much space
on the left looks bad. What I should actually be doing is to move
the AnchorJS links to the right when there is insufficient space on
the left. Media queries are probably the best way to do that. This
commit doesn't do anything like that.

I don't plan to keep it this way.