restore original padding (roughly speaking)
The left and right side padding were 10px before I started messing
with it to make the AnchorJS links always appear. I've made it 1rem,
which, given the root font-size, is 10px at default scaling.

I'll likely reduce it further, but I want to have a commit with
this, since on *most* pages this amount of padding looks good.