moved AnchorJS links to the right
I think these are less convenient and look worse than on the left,
but I don't want to take the time now to figure out [a] how to make
them move responsively to the right when the viewport is narrow
enough that they are cut off on the left (this appears nontrivial
with AnchorJS, which don't add links on both sides), and [b] whether
that would itself introduce additional confusion.

I've adjusted the links' colors accordingly, so big ones aren't too
obtrusive, while small ones are better noticed.