set pandoc input format to GitHub Flavored Markdown
This fixes most of the heading ids. But not the ones that contain
<br>s. GitHub doesn't put anything for the <br>; but pandoc, even
with gfm, puts a single hyphen. Adding a space next to the <br>
causes GitHub to put a hyphen... but causes pandoc (set to gfm) to
put two consecutive hyphens.

The least bad way to fix that might end up being to write a pandoc
filter in Lua.

Made a couple related tweaks: an &rsquo; broke for some reason, and
the hack that fixed it was also better than a span in another place
for preventing a linkification.