docs tree restructure
I'm writing custom pandoc filters in Lua anyway (see recent
commits) so I might as well keep all the documentation, other than
README.md (which also serves as the repo readme), in a doc dir.
Stripping "doc/" prefixes in anchor hrefs is simple with a filter.

This reverses course on the decision in 54d571d.

doc-assets is renamed to doc and README.html resides there instead
of in the repo root. Other files than README.md that are used by
pandoc (such that if they change then README.html should be
regenerated) are in doc/pandoc.

I may end up renaming README.html back to index.html, but I don't
want to do that in this commit, because I want to make it possible
to revert that separately in the future and to see easily which
specific changes (since that will change multiple files, too) are
associated with it.