reworded to avoid making it sound like SharpAvi wraps toasts too
The use of commas around "which wraps VfW" may have been sufficient
for readers accustomed to American English (I'm not sure), but some
other dialects don't use commas so strictly, so I think the risk of
confusion, for a point that is so important, is too high.

I still hope to find a better wording, though.