trying stickyNavbar.js
Maybe it will work better if the navbar is styled properly (as a
bar), but I don't think that can adequately explain how low it is
kept.

Even if I fix this, it seems like it might involve a comparable
amount of code and work to making a sticky navbar manually,
especially since its documentation suggests that it might not work
well on mobile browsers (though I don't know why that would be).