made navlinks smooth-scroll (and not put #id in URL) on primary click
This uses a polyfill to get smooth scrolling in IE, which I've
added to the Dependencies and Notices sections.

Smooth scrolling, for this purpose, was the last thing I'd feared I
might need jQuery for. So now that this is implemented, it makes
sense to rewrite the one function that uses jQuery and remove the
dependency. (This commit does not do that, though.)