added ../ prefix to local links outside doc/
This fixes a bug introduced in 189cedf, adding the other half of
the job that the Link filter there did (as well as some error
checking for links that would need to be adjusted but that the
existing logic doesn't cover). Now links to files like LICENSE
work.

It would be sufficient to *always* prefix ../ and never remove
doc/, since, when in doc, ../doc/X goes to X. But I think it's
worth having a function that's a little more complicated but
produces results that are more aesthetically pleasing and less
confusing to look at (on hover, browsers will show them the same,
but one may see the difference under other circumstances, as when
examining the HTML, examining the DOM, or in the text of error
messages).