started on making links work properly (most externally)
I need to solve the problem of how to open a file of an arbitrary
type and name in the default web browser. ShellExecute doesn't
support this naturally.

Unfortunately, ShellExecute also doesn't support passing along
URL query parameters, so the current attempt, involving a separate
special page that uses JavaScript to check its query string and do
the redirection, cannot work.

I'm committing in this state in case some of the code I wrote to do
that ever needs to be looked at later. After this, I'm going to rip
that code out and do something different.

Note that this problem is only part of the changes made here. Many
of the changes here are good and will remain.