borders and "shadows" in rem instead of px for better scaling
This is partiuclarly significant at high resolutions with high
display scaling. For example, at 3840x2160 with 350% display
scaling, the "shadow" on the currently active section navlink is
only *just* noticeable (when the page is viewed in the WebBrowser
control, which doesn't scale px, at least not by default). When
rem units are used instead (with values equivalent to the old px
units at 100% scaling), the "shadow" is fully visible (with the
proper proportions).