added menu shortcut keys for all menu items
NB these are not the same as what I confusingly called "shortcut
keys" in the README (now also fixed). What I've added here are just
shortcut keys that one presses while the menu is open (and that are
revealed by opening the menu with Alt+Space instead of the mouse).

Each menu item should also eventually have a keyboard shortcut
(that is, an accelerator): a combination of keys that, when pressed
at any time while the main application window has focus, activates
it (as if it were clicked in the menu).

But I may have to make some other changes to do that without making
the UI more, rather than less, confusing. See "Menu items are
cumbersome to access while drawing" and "It might be useful to
pause" in the README. (That part of the README is not new; the only
change to it in this commit is fixing a confusing word choice.)