converted reusasble menu-manipulation to use composition, not inheritance
This still runs fine. It also works with the designer in VS2019
16.10.10, the current stable version. It does *not* work with the
designer in VS2019 16.11.0 Preview 1.0, and the error message there
is uninformative. I believe this is due to a bug in the preview
version. Hopefully that bug will be fixed. It might be good to try
and produce a simpler code sample that triggers the same problem
and file a bug report. I don't want this to break in the designer
due to an update.

This refactor itself was mainly motivated by the goal of being
compatible with the designer, though it has some other benefits.

I did this since it turned out to be very complicated to get the
inheritance way (in which MainWindow, and eventually HelpWindow,
would each inherit from an abstract constructed type that would
themselves each inherit from an abstact constructed type) to be
compatible with the Windows Forms designer, which needs special
work to display/design forms that inherit from classes that are
abstract, generic, or both:

https://web.archive.org/web/20200811092319/http://www.pocketsilicon.com/post/Using-Visual-Studio-Whidbey-to-Design-Abstract-Forms

I tried but did not manage to get such a technique to work. But
even if I had, I'd likely abandon it and reimplement a composition-
based way (as I've done in this commit) to avoid carrying and
maintaining a whole bunch of complex, esoteric code that
conceptually has nothing to do with VidDraw's functionality or user
interface.

Since C# doesn't have private inheritance or mutiple inheritance,
really composition should be used for this kind of "implemented in
terms of" relationship anyway. But that's not the main reason I've
done this.