optimize updateActiveNavlink to only touch the nav if the section has changed
This would be a worthwhile improvement in and of itself, since it
buys improved performance at the cost of only slightly increased
complexity (also, the code is more intuitive now, so arguably this
would be worthwhile even if it never improved performance).

But the main reason I'm doing this is to prepare to extend it so
that, when the page is being viewed in VidDraw's help viewer, the
page tells the help viewer about section changes, so the help
viewer can update check-mark placement on system-menu items for
help sections. I don't want to update the system menu many times
more than necessary, since that could make it harder for users to
smoothly open it to do things like closing the window... which
could make for a very unpleasant user experience. (I also expect
that performance would suffer due to all the marshaling.)