put licenses in Markdown (but not HTML) readme in <details> elements
This is to cut down on the excessive length of README.md as it
appears on GitHub, where the long Apache License is not vertically
bounded and scolled.

I don't think this is better for index.html, so I've use a Pandoc
filter to remove it.

This commit has some small related rewording, <summary> elements
inside the <details> elements in the Markdown version, and comments
added to all the local functions in the Pandoc filter (including
those unrelated to those changes).