RawSystemMenu helpers leave unrelated fState bits alone
This is technically a refactor, in the sense that the helpers were
being used in a way that kept this bug from ever affecting the
behavior of the program, though the situation was brittle.

Fixing this bug made it easy to remove the declaration of
EnableMenuItem and implement SetEnabled in analogously to SetCheck.
If I had done this before, then enabling or graying a checked menu
item would remove the check mark. (That's how I found the bug.)