link manpages in latest Ubuntu LTS
The links I had just switched them to before are actually 404 pages
that manage to redirect to current documentation:

https://github.com/EliahKagan/VidDraw/actions/runs/3920635091/jobs/6702377963

This reminded me that those redirects have not always worked, and
I'm not sure Canonical actually intends that we rely them. At least
for now, I'm going back to linking to specific releases' manpages
(but current releases, so the links resolve properly).

I'm not really sure what the best thing is to do here, since
arguably the 404 links that redirect are more reliable, in the
sense that releases merely going end-of-life is insufficient to
break them.