Experiment with initializeCommand
The ultimate goal is to figure out how git is configured for line
endings on the host that is creating the Docker container, and pass
information about this into the container, so the autoCrlf setting
the repository was cloned with on the host will be used by git in
the container.

But for now, I just want to see what logs will show the output of
an initializeCommand.