Carry in config when repo is *directly* in Windows
This specific situation was not covered before.

This now covers the hopefully infrequent case where the user
reopens the local directory, in a Windows system, in the container,
rather than cloning the repository in the container.

In that situation, initializeCommand runs *directly* on the host
Windows system. In contrast, in other situations on Windows, it
does not run *directly* on Windows, but instead in the Linux-based
environment that Docker uses do its work. (This is currently either
a WSL 2 system or a Linux-based system running directly in a
Hyper-V virtual machine, depending which backend Docker is
configured to use.)

To cover this specific situation, I expand the noop stub added in
2bba978, .devcontainer/initialize.cmd, so it does the same work as
.devcontainer/initialize. initialize.cmd is now a translation of
initialize from bash to the batch language.

This commit also makes some stylistic improvements in the bash
scripts.